simplify hydro remix : object Storage becomes a blank parameter of the algorithm shavePeaksByRemixingHydro(...)
