+++ src/tests/src/solver/simulation/test-hydro-remix.cpp
@@ -3,6 +3,7 @@
 #define WIN32_LEAN_AND_MEAN
 
 #include <algorithm>
+#include <memory>
 #include <unit_test_utils.h>
 #include <vector>
 
@@ -27,10 +28,22 @@ struct InputFixture
         pump.assign(size, 0.);
         Spillage.assign(size, 0.);
         DTG_MRG.assign(size, 0.);
+        hydroStorage = std::make_shared<HydroStorage>(HydroGen,
+                                                      UnsupE,
+                                                      HydroPmax,
+                                                      HydroPmin,
+                                                      inflows,
+                                                      ovf,
+                                                      pump,
+                                                      init_level,
+                                                      capacity,
+                                                      pumpEff,
+                                                      reservoirManagement);
     }
 
     std::vector<double> TotalGenNoHydro, HydroGen, UnsupE, HydroPmax, HydroPmin, inflows, ovf, pump,
       Spillage, DTG_MRG;
+    std::shared_ptr<Storage> hydroStorage;
     double init_level = 0.;
     double capacity = std::numeric_limits<double>::max();
     const double pumpEff = 1.0;
@@ -40,7 +53,7 @@ struct InputFixture
 BOOST_FIXTURE_TEST_CASE(input_vectors_of_different_sizes__exception_raised, InputFixture<1>)
 {
     HydroGen = {0., 0.};
-
+    
     BOOST_CHECK_EXCEPTION(shavePeaksByRemixingHydro(HydroGen,
                                                     UnsupE,
                                                     TotalGenNoHydro,
@@ -54,7 +67,8 @@ BOOST_FIXTURE_TEST_CASE(input_vectors_of_different_sizes__exception_raised, Inpu
                                                     ovf,
                                                     pump,
                                                     Spillage,
-                                                    DTG_MRG),
+                                                    DTG_MRG,
+                                                    hydroStorage),
                           std::invalid_argument,
                           checkMessage("Remix hydro input : arrays of different sizes"));
 }
@@ -77,7 +91,8 @@ BOOST_FIXTURE_TEST_CASE(input_init_level_exceeds_capacity__exception_raised, Inp
                                                     ovf,
                                                     pump,
                                                     Spillage,
-                                                    DTG_MRG),
+                                                    DTG_MRG,
+                                                    hydroStorage),
                           std::invalid_argument,
                           checkMessage("Remix hydro input : initial level > reservoir capacity"));
 }
@@ -100,7 +115,8 @@ BOOST_FIXTURE_TEST_CASE(all_input_arrays_of_size_0__exception_raised, InputFixtu
                                                     ovf,
                                                     pump,
                                                     Spillage,
-                                                    DTG_MRG),
+                                                    DTG_MRG,
+                                                    hydroStorage),
                           std::invalid_argument,
                           checkMessage("Remix hydro input : all arrays of sizes 0"));
 }
@@ -126,7 +142,8 @@ BOOST_FIXTURE_TEST_CASE(Hydro_gen_not_smaller_than_pmax__exception_raised, Input
                                 ovf,
                                 pump,
                                 Spillage,
-                                DTG_MRG),
+                                DTG_MRG,
+                                hydroStorage),
       std::invalid_argument,
       checkMessage("Remix hydro input : Hydro generation not smaller than Pmax everywhere"));
 }
@@ -152,7 +169,8 @@ BOOST_FIXTURE_TEST_CASE(Hydro_gen_not_greater_than_pmin__exception_raised, Input
                                 ovf,
                                 pump,
                                 Spillage,
-                                DTG_MRG),
+                                DTG_MRG,
+                                hydroStorage),
       std::invalid_argument,
       checkMessage("Remix hydro input : Hydro generation not greater than Pmin everywhere"));
 }
@@ -175,7 +193,8 @@ BOOST_FIXTURE_TEST_CASE(input_is_acceptable__no_exception_raised, InputFixture<1
                                                    ovf,
                                                    pump,
                                                    Spillage,
-                                                   DTG_MRG));
+                                                   DTG_MRG,
+                                                   hydroStorage));
 }
 
 BOOST_FIXTURE_TEST_CASE(
@@ -202,7 +221,8 @@ BOOST_FIXTURE_TEST_CASE(
                               ovf,
                               pump,
                               Spillage,
-                              DTG_MRG);
+                              DTG_MRG,
+                              hydroStorage);
 
     std::vector<double> expected_HydroGen = {20., 20., 20., 20., 20.};
     // UnsupE such as TotalGenNoHydro + HydroGen + UnsupE remains flat
@@ -233,7 +253,8 @@ BOOST_FIXTURE_TEST_CASE(Pmax_does_not_impact_results_when_greater_than_40mwh, In
                               ovf,
                               pump,
                               Spillage,
-                              DTG_MRG);
+                              DTG_MRG,
+                              hydroStorage);
 
     std::vector<double> expected_HydroGen = {20., 20., 20., 20., 20.};
     // UnsupE such as TotalGenNoHydro + HydroGen + UnsupE remains constant at each hour
@@ -266,7 +287,8 @@ BOOST_FIXTURE_TEST_CASE(
                               ovf,
                               pump,
                               Spillage,
-                              DTG_MRG);
+                              DTG_MRG,
+                              hydroStorage);
 
     std::vector<double> expected_HydroGen = {20., 20., 20., 20., 20.};
     // UnsupE such as TotalGenNoHydro + HydroGen + UnsupE remains constant at each hour
@@ -301,7 +323,8 @@ BOOST_FIXTURE_TEST_CASE(influence_of_pmax, InputFixture<5>, *boost::unit_test::t
                               ovf,
                               pump,
                               Spillage,
-                              DTG_MRG);
+                              DTG_MRG,
+                              hydroStorage);
 
     std::vector<double> expected_HydroGen_1 = {0., 0., 13.33, 33.33, 53.33};
     BOOST_TEST(HydroGen == expected_HydroGen_1, boost::test_tools::per_element());
@@ -324,7 +347,8 @@ BOOST_FIXTURE_TEST_CASE(influence_of_pmax, InputFixture<5>, *boost::unit_test::t
                               ovf,
                               pump,
                               Spillage,
-                              DTG_MRG);
+                              DTG_MRG,
+                              hydroStorage);
 
     std::vector<double> expected_HydroGen_2 = {20., 20., 20., 20., 20.};
     std::vector<double> expected_UnsupE_2 = {50., 50., 50., 50., 50.};
@@ -358,7 +382,8 @@ BOOST_FIXTURE_TEST_CASE(influence_of_pmin, InputFixture<5>, *boost::unit_test::t
                               ovf,
                               pump,
                               Spillage,
-                              DTG_MRG);
+                              DTG_MRG,
+                              hydroStorage);
     std::vector<double> expected_HydroGen_1 = {0., 0., 13.33, 33.33, 53.33};
     BOOST_TEST(HydroGen == expected_HydroGen_1, boost::test_tools::per_element());
 
@@ -379,7 +404,8 @@ BOOST_FIXTURE_TEST_CASE(influence_of_pmin, InputFixture<5>, *boost::unit_test::t
                               ovf,
                               pump,
                               Spillage,
-                              DTG_MRG);
+                              DTG_MRG,
+                              hydroStorage);
 
     std::vector<double> expected_HydroGen_2 = {20., 20., 20., 20., 20.};
     std::vector<double> expected_UnsupE_2 = {50., 50., 50., 50., 50.};
@@ -409,7 +435,8 @@ BOOST_FIXTURE_TEST_CASE(Hydro_gen_is_already_flat___remix_is_useless__level_easi
                                             ovf,
                                             pump,
                                             Spillage,
-                                            DTG_MRG);
+                                            DTG_MRG,
+                                            hydroStorage);
 
     std::vector<double> expected_levels = {480., 460., 440., 420., 400.};
     BOOST_TEST(levels == expected_levels, boost::test_tools::per_element());
@@ -438,7 +465,8 @@ BOOST_FIXTURE_TEST_CASE(input_leads_to_levels_over_capacity___exception_raised, 
                                 ovf,
                                 pump,
                                 Spillage,
-                                DTG_MRG),
+                                DTG_MRG,
+                                hydroStorage),
       std::invalid_argument,
       checkMessage(
         "Remix hydro input : levels computed from input don't respect reservoir bounds"));
@@ -466,7 +494,8 @@ BOOST_FIXTURE_TEST_CASE(input_leads_to_levels_less_than_zero___exception_raised,
                                 ovf,
                                 pump,
                                 Spillage,
-                                DTG_MRG),
+                                DTG_MRG,
+                                hydroStorage),
       std::invalid_argument,
       checkMessage(
         "Remix hydro input : levels computed from input don't respect reservoir bounds"));
@@ -503,7 +532,8 @@ BOOST_FIXTURE_TEST_CASE(influence_of_capacity_on_algorithm___case_where_no_influ
                                             ovf,
                                             pump,
                                             Spillage,
-                                            DTG_MRG);
+                                            DTG_MRG,
+                                            hydroStorage);
 
     // HydroGen is flat and is 15. (means of initial HydroGen)
     std::vector<double> expected_HydroGen(10, 15.);
@@ -531,7 +561,8 @@ BOOST_FIXTURE_TEST_CASE(influence_of_capacity_on_algorithm___case_where_no_influ
                                        ovf,
                                        pump,
                                        Spillage,
-                                       DTG_MRG);
+                                       DTG_MRG,
+                                       hydroStorage);
 
     BOOST_TEST(HydroGen == expected_HydroGen, boost::test_tools::per_element());
     BOOST_TEST(levels == expected_L, boost::test_tools::per_element());
@@ -568,7 +599,8 @@ BOOST_FIXTURE_TEST_CASE(lowering_capacity_too_low_leads_to_suboptimal_solution_f
                                             ovf,
                                             pump,
                                             Spillage,
-                                            DTG_MRG);
+                                            DTG_MRG,
+                                            hydroStorage);
 
     // HydroGen is flat and is 15. (means of initial HydroGen)
     std::vector<double> expected_HydroGen(10, 15.);
@@ -599,7 +631,8 @@ BOOST_FIXTURE_TEST_CASE(lowering_capacity_too_low_leads_to_suboptimal_solution_f
                                        ovf,
                                        pump,
                                        Spillage,
-                                       DTG_MRG);
+                                       DTG_MRG,
+                                       hydroStorage);
 
     // OutHydroGen_2 is flat by interval
     std::vector<double> expected_HydroGen_2 = {16., 16., 16., 16., 16., 14., 14., 14., 14., 14.};
@@ -638,7 +671,8 @@ BOOST_FIXTURE_TEST_CASE(lowering_initial_level_too_low_leads_to_suboptimal_solut
                                             ovf,
                                             pump,
                                             Spillage,
-                                            DTG_MRG);
+                                            DTG_MRG,
+                                            hydroStorage);
 
     // HydroGen is flat and is 25. (means of initial HydroGen)
     std::vector<double> expected_HydroGen(10, 25.);
@@ -667,7 +701,8 @@ BOOST_FIXTURE_TEST_CASE(lowering_initial_level_too_low_leads_to_suboptimal_solut
                               ovf,
                               pump,
                               Spillage,
-                              DTG_MRG);
+                              DTG_MRG,
+                              hydroStorage);
 
     // OutHydroGen_2 is flat by interval
     std::vector<double> expected_HydroGen_2 = {24., 24., 24., 24., 24., 26., 26., 26., 26., 26.};
@@ -704,7 +739,8 @@ BOOST_FIXTURE_TEST_CASE(influence_of_initial_level_on_algorithm___case_where_no_
                                             ovf,
                                             pump,
                                             Spillage,
-                                            DTG_MRG);
+                                            DTG_MRG,
+                                            hydroStorage);
 
     // HydroGen is flat and is 15. (means of initial HydroGen)
     std::vector<double> expected_HydroGen(10, 15.);
@@ -733,7 +769,8 @@ BOOST_FIXTURE_TEST_CASE(influence_of_initial_level_on_algorithm___case_where_no_
                                        ovf,
                                        pump,
                                        Spillage,
-                                       DTG_MRG);
+                                       DTG_MRG,
+                                       hydroStorage);
 
     // OutHydroGen_2 is flat (and optimal)
     std::vector<double> expected_HydroGen_2(10, 15.);
@@ -765,7 +802,8 @@ BOOST_FIXTURE_TEST_CASE(spillage_positive_at_hour_0___no_change_at_this_hour, In
                               ovf,
                               pump,
                               Spillage,
-                              DTG_MRG);
+                              DTG_MRG,
+                              hydroStorage);
 
     std::vector<double> expected_HydroGen = {40., 15., 15., 15., 15.};
     BOOST_CHECK(HydroGen == expected_HydroGen);
@@ -796,7 +834,8 @@ BOOST_FIXTURE_TEST_CASE(DTG_MRG_positive_on_hour_4___no_change_at_this_hour, Inp
                               ovf,
                               pump,
                               Spillage,
-                              DTG_MRG);
+                              DTG_MRG,
+                              hydroStorage);
 
     std::vector<double> expected_HydroGen = {25., 25., 25., 25., 0.};
     BOOST_CHECK(HydroGen == expected_HydroGen);
@@ -837,7 +876,8 @@ BOOST_FIXTURE_TEST_CASE(comparison_of_results_with_python_algo,
                               ovf,
                               pump,
                               Spillage,
-                              DTG_MRG);
+                              DTG_MRG,
+                              hydroStorage);
     std::vector<double> expected_HydroGen = {42.3, 35.3,  27.,  31.,   7.,    33.,   8.,
                                              31.,  19.55, 2.,   38.55, 30.55, 22.55, 7.,
                                              4.,   45.55, 6.55, 25.55, 41.55, 25.};
